requirements.txt
Flask
requests
openai
azure-cosmos
